Add files via upload
Files for my final year project including my viva and feedback for said viva. I was very proud of this project which I conducted independently from no prior knowledge on neural nets. My overall grade for this dissertation module (investigation in mathematics) was 94% overall with the final report and viva accounting for 90% of the final grade.